requirements.txt
tkinter
websocket-client
requests
numpy
matplotlib